+++ src/scout_apm/django/middleware.py
@@ -12,7 +12,6 @@ from scout_apm.core.tracked_request import TrackedRequest
 from scout_apm.core.web_requests import (
     create_filtered_path,
     ignore_path,
-    track_amazon_request_queue_time,
     track_request_queue_time,
 )
 
@@ -168,11 +167,7 @@ class MiddlewareTimingMiddleware(object):
         queue_time = request.META.get("HTTP_X_QUEUE_START") or request.META.get(
             "HTTP_X_REQUEST_START", ""
         )
-        queue_time_tracked = track_request_queue_time(queue_time, tracked_request)
-        if not queue_time_tracked:
-            track_amazon_request_queue_time(
-                request.META.get("HTTP_X_AMZN_TRACE_ID", ""), tracked_request
-            )
+        track_request_queue_time(queue_time, tracked_request)
 
         with tracked_request.span(
             operation="Middleware",
@@ -254,11 +249,7 @@ class OldStyleMiddlewareTimingMiddleware(object):
         queue_time = request.META.get("HTTP_X_QUEUE_START") or request.META.get(
             "HTTP_X_REQUEST_START", ""
         )
-        queue_time_tracked = track_request_queue_time(queue_time, tracked_request)
-        if not queue_time_tracked:
-            track_amazon_request_queue_time(
-                request.META.get("HTTP_X_AMZN_TRACE_ID", ""), tracked_request
-            )
+        track_request_queue_time(queue_time, tracked_request)
 
         tracked_request.start_span(
             operation="Middleware", should_capture_backtrace=False