+++ src/scout_apm/pyramid.py
@@ -7,7 +7,6 @@ from scout_apm.core.tracked_request import TrackedRequest
 from scout_apm.core.web_requests import (
     create_filtered_path,
     ignore_path,
-    track_amazon_request_queue_time,
     track_request_queue_time,
 )
 
@@ -53,14 +52,10 @@ def instruments(handler, registry):
                 )
                 tracked_request.tag("user_ip", user_ip)
 
-            tracked_queue_time = False
             queue_time = request.headers.get(
                 "x-queue-start", default=""
             ) or request.headers.get("x-request-start", default="")
-            tracked_queue_time = track_request_queue_time(queue_time, tracked_request)
-            if not tracked_queue_time:
-                amazon_queue_time = request.headers.get("x-amzn-trace-id", default="")
-                track_amazon_request_queue_time(amazon_queue_time, tracked_request)
+            track_request_queue_time(queue_time, tracked_request)
 
             try:
                 try: